remove(proxy): delete lifecycle tests, reduce unused middleware logging, and simplify private flag description
